+++ aasrp/static/aasrp/javascript/aa-srp-dashboard.js
@@ -266,63 +266,62 @@ $(document).ready(function () {
      */
     // enable link modal
     $('#enable-srp-link').on('show.bs.modal', function (event) {
-        let button = $(event.relatedTarget);
-        let url = button.data('url');
-        let name = button.data('name');
-        let modal = $(this);
+        const button = $(event.relatedTarget);
+        const url = button.data('url');
+        const name = button.data('name');
+        const modal = $(this);
 
         modal.find('#modal-button-confirm-enable-srp-link').attr('href', url);
         modal.find('.modal-body').html(
             aaSrpSettings.translation.modal.enableSrpLink.body + '<br>"' + name + '"'
         );
     }).on('hide.bs.modal', function () {
-        let modal = $(this);
+        const modal = $(this);
 
         modal.find('.modal-body').html('');
     });
 
     // disable link modal
     $('#disable-srp-link').on('show.bs.modal', function (event) {
-        let button = $(event.relatedTarget);
-        let url = button.data('url');
-        let name = button.data('name');
-        let modal = $(this);
+        const button = $(event.relatedTarget);
+        const url = button.data('url');
+        const name = button.data('name');
+        const modal = $(this);
 
         modal.find('#modal-button-confirm-disable-srp-link').attr('href', url);
         modal.find('.modal-body').html(
             aaSrpSettings.translation.modal.disableSrpLink.body + '<br>"' + name + '"'
         );
     }).on('hide.bs.modal', function () {
-        let modal = $(this);
+        const modal = $(this);
 
         modal.find('.modal-body').html('');
     });
 
     // delete link modal
     $('#delete-srp-link').on('show.bs.modal', function (event) {
-        let button = $(event.relatedTarget);
-        let url = button.data('url');
-        let name = button.data('name');
-        let modal = $(this);
+        const button = $(event.relatedTarget);
+        const url = button.data('url');
+        const name = button.data('name');
+        const modal = $(this);
 
         modal.find('#modal-button-confirm-delete-srp-link').attr('href', url);
         modal.find('.modal-body').html(
             aaSrpSettings.translation.modal.deleteSrpLink.body + '<br>"' + name + '"'
         );
     }).on('hide.bs.modal', function () {
-        let modal = $(this);
+        const modal = $(this);
 
         modal.find('.modal-body').html('');
     });
 
     // show details
     $('#srp-request-details').on('show.bs.modal', function (event) {
-        let button = $(event.relatedTarget);
-        let modal = $(this);
-
-        let name = button.data('modal-title');
-        let url = button.data('link');
-        let confirmButtonText = button.data('modal-button-confirm');
+        const button = $(event.relatedTarget);
+        const modal = $(this);
+        const name = button.data('modal-title');
+        const url = button.data('link');
+        const confirmButtonText = button.data('modal-button-confirm');
 
         modal.find('.modal-title').text(name);
         modal.find('#modal-button-request-details-confirm').html(confirmButtonText);
@@ -360,7 +359,7 @@ $(document).ready(function () {
             }
         });
     }).on('hide.bs.modal', function () {
-        let modal = $(this);
+        const modal = $(this);
 
         modal.find('.modal-title').text('');
         modal.find('.modal-body').text('');